sql_db/requirements.txt
Flask
mysql-connector-python
pandas
SQLAlchemy
```